place star rating near like button
